refactor(v0.85.0): cli/_helpers IO/key utilities → core/utils/env_io.py 추출 (A5a)
A5 deferred 작업 재개의 첫 번째 PR. v0.82.0 점검에서 startup.py를
core/lifecycle/로 옮길 때 cli._helpers 의존이 layer 위반을 새로 만들어
A5 폐기됐던 결함 해소.

Strategy: A5a (this PR) → _helpers의 4개 IO/key util을 utils로 추출.
A5b (next PR) → startup.py 책임 분리.

이전 (cli/_helpers.py 113L → 21L):
- mask_key, upsert_env, upsert_config_toml, is_glm_key (107L) → core/utils/env_io.py
- parse_dry_run_flag (15L) → cli/_helpers에 유지 (진짜 CLI argument 파싱)

호출자 5 파일 업데이트:
- core/cli/startup.py: mask_key/upsert_env/is_glm_key 3 imports
- core/cli/commands/__init__.py: 3 imports
- core/cli/commands/model.py: upsert_config_toml 1 import
- tests/test_config_effort_knob.py: upsert_config_toml 1 import
- core/cli/dispatcher.py: parse_dry_run_flag 그대로 (변경 없음)

ruff auto-fix: 2 import-sort 자동 정리.

Quality Gates:
- ruff/mypy 0 errors (331 source files, +1)
- pytest 4345 passed, 1 skipped, 24 deselected — 변동 없음
- import-linter 4/4 contracts kept (still)
- E2E Cowboy Bebop A (68.4) 변동 없음

다음 PR (A5b): cli/startup.py를 core/lifecycle/startup.py로 이전 +
interactive wizard 부분은 core/cli/onboarding.py로 분리.
